fix(@tyeps/mongodb): reorganize tests into different files (old test are mostly untouched)
