+++ types/mongodb/tsconfig.json
@@ -18,6 +18,16 @@
     },
     "files": [
         "index.d.ts",
-        "mongodb-tests.ts"
+        "test/index.ts",
+        "test/transaction.ts",
+        "test/stats.ts",
+        "test/cursor.ts",
+        "test/createIndex.ts",
+        "test/collection/findX.ts",
+        "test/collection/insertX.ts",
+        "test/collection/updateX.ts",
+        "test/collection/mapReduce.ts",
+        "test/collection/aggregate.ts",
+        "test/collection/count.ts"
     ]
 }
